Move DockerCompose to external repo
