+++ test-helpers/src/docker_compose.rs
@@ -1,6 +1,8 @@
-use crate::docker_compose_runner::*;
+use docker_compose_runner::*;
 use std::{env, path::Path};
 use tracing_subscriber::fmt::TestWriter;
+
+pub use docker_compose_runner::DockerCompose;
 
 fn setup_tracing_subscriber_for_test_logic() {
     tracing_subscriber::fmt()
@@ -77,7 +79,7 @@ fn build_images(service_to_image: &[&str]) {
         .iter()
         .any(|x| *x == "shotover-int-tests/cassandra:4.0.6")
     {
-        run_command(
+        crate::run_command(
             "docker",
             &[
                 "build",
@@ -92,7 +94,7 @@ fn build_images(service_to_image: &[&str]) {
         .iter()
         .any(|x| *x == "shotover-int-tests/cassandra:3.11.13")
     {
-        run_command(
+        crate::run_command(
             "docker",
             &[
                 "build",
@@ -109,7 +111,7 @@ fn build_images(service_to_image: &[&str]) {
         && Path::new("example-configs/docker-images/cassandra-tls-4.0.6/certs/keystore.p12")
             .exists()
     {
-        run_command(
+        crate::run_command(
             "docker",
             &[
                 "build",
